requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
aws-lambda-builders==1.50.0
aws-lambda-powertools==3.4.1
aws-psycopg2==1.3.8
aws-sam-cli==1.122.0
aws-sam-translator==1.90.0
aws-xray-sdk==2.14.0
awscrt==0.24.1
awsiotsdk==1.0.0.dev0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.138
boto3-stubs==1.34.159
botocore==1.34.138
botocore-stubs==1.35.0
Brlapi==0.7.0
cachetools==5.5.2
catkin-pkg==0.4.16
cdifflib==1.2.6
certifi==2024.6.2
cffi==1.17.0
cfn-lint==1.10.3
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
contourpy==1.1.1
cookiecutter==2.6.0
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.0
dbus-python==1.2.16
deepdiff==7.0.1
defer==1.0.6
dell-recovery==0.0.0
Deprecated==1.2.14
deskew==1.4.3
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.4
django-admin-tools==0.9.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.6.1
docker==7.1.0
docutils==0.15.2
docx2txt==0.8
duplicity==0.8.12.0
email-validator==2.1.2
emoji==2.12.1
empy==3.3.2
entrypoints==0.3
epub2pdf==0.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.14.1
filelock==3.15.3
filetype==1.2.0
fire==0.6.0
flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
future==0.18.2
gencpp==0.6.2
genlisp==0.4.17
genmsg==0.5.14
genpy==0.6.9
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.35.0
img2pdf==0.5.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
imutils==0.5.4
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonfield==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==18.0.1
kiwisolver==1.4.5
langchain==0.2.6
langchain-aws==0.1.11
langchain-community==0.2.6
langchain-core==0.2.22
langchain-experimental==0.0.62
langchain-postgres==0.0.9
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.79
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.4
lockfile==0.12.2
loguru==0.6.0
louis==3.12.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mobi==0.3.3
monotonic==1.5
mpmath==1.3.0
multidict==6.0.5
mypy-boto3-apigateway==1.34.137
mypy-boto3-cloudformation==1.34.111
mypy-boto3-ecr==1.34.154
mypy-boto3-iam==1.34.162
mypy-boto3-kinesis==1.34.153
mypy-boto3-lambda==1.34.77
mypy-boto3-s3==1.34.162
mypy-boto3-schemas==1.34.0
mypy-boto3-secretsmanager==1.34.145
mypy-boto3-signer==1.34.95
mypy-boto3-sqs==1.34.121
mypy-boto3-stepfunctions==1.34.149
mypy-boto3-sts==1.34.0
mypy-boto3-xray==1.34.0
mypy-extensions==1.0.0
narwhals==1.42.1
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opensearch-py==2.3.1
ordered-set==4.1.0
orjson==3.10.5
packaging==24.1
pandas==2.0.3
paramiko==2.6.0
passlib==1.7.4
pdf2docx==0.5.8
pdf2image==1.17.0
pdfkit==1.0.0
pexpect==4.6.0
pgvector==0.2.5
pikepdf==8.15.1
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
plotly==6.3.0
progressbar==2.5
propcache==0.2.0
protobuf==5.29.5
psutil==5.5.1
psycopg==3.1.19
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pycairo==1.16.2
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.20.0
pycryptodomex==3.20.0
pycups==1.9.73
pydantic==2.7.4
pydantic-core==2.18.4
pydeck==0.9.1
pygments==2.18.0
PyGObject==3.36.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyNaCl==1.3.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypdf==4.2.0
pypdf2==3.0.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2023.3
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==6.0.1
rapidfuzz==3.9.4
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
reportlab==3.5.34
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rich==13.7.1
roman==2.0.0
rospkg==1.2.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
shellingham==1.5.4
simplejson==3.16.0
sip==4.19.21
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sos==4.8.2
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
starlette==0.37.2
streamlit==1.40.1
sympy==1.13.2
systemd-python==234
tabulate==0.9.0
tenacity==8.4.1
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
tomlkit==0.13.0
tornado==6.4.2
tqdm==4.66.4
typer==0.12.3
types-awscrt==0.21.2
types-python-dateutil==2.9.0.20240316
types-s3transfer==0.10.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ujson==5.10.0
unattended-upgrades==0.1
unstructured==0.11.8
unstructured-client==0.24.1
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
wadllib==1.3.3
watchdog==4.0.2
watchfiles==0.22.0
websockets==12.0
werkzeug==3.0.3
wkhtmltopdf==0.2
wrapt==1.16.0
xkit==0.0.0
yarl==1.9.4
zipp==3.19.2
